syntax: simplify hir::GroupKind
I'm not sure exactly why I used three variants instead of two like how
I've defined it in this patch. Possibly because the AST uses three
variants? (The AST needs to do a little more work to store a span
associated with where the name actually is in the expression, so it
maybe makes a little more sense there.)

In any case, this is the first step of many in simplifying the HIR.