syntax: improve Debug impls
This makes the Debug impls for Literal and ClassRangeBytes a bit better.
The former in particular. Instead of just printing a sequence of decimal
numbers, we now print them as characters.

Given the lackluster support for Vec<u8> as a string in the standard
library, we copy a little bit of code from regex-automata to make the
debug print for the Vec<u8> basically as nice as a String.