syntax: rewrite 'cls1|..|clsN' as '[cls1..clsN]'
Whenever we have an alternation where each of its branches are just
classes, we can always combined that into a single class. Single classes
are generally going to be cheaper to process further down the pipeline.
Namely, instead of needing to branch between them at a higher level in
an NFA graph, they can handled as one single unit.